+++ src/main/java/org/example/uberbookingservice/repositories/bookingRepository.java
@@ -3,5 +3,5 @@ package org.example.uberbookingservice.repositories;
 import com.example.EntityService.Models.Booking;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface bookingRepository extends JpaRepository<Booking,Long> {
+public interface bookingRepository extends JpaRepository<Booking,Integer> {
 }
